requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
croniter==2.0.5
discord-py-interactions==5.12.1
discord-typings==0.9.0
emoji==2.12.1
frozenlist==1.4.1
idna==3.7
multidict==6.0.5
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
tomli==2.0.1
typing_extensions==4.12.2
yarl==1.9.4
